ISSUE #1916 - Move all methods to one common class
