+++ frontend/services/viewer/viewerService/viewerController.tsx
@@ -15,20 +15,21 @@
  **  along with this program.  If not, see <http=//www.gnu.org/licenses/>.
  **/
 
+import EventEmitter from 'eventemitter3';
 import React from 'react';
+
 import { IS_DEVELOPMENT } from '../../../constants/environment';
 import {
 	VIEWER_EVENTS,
 	VIEWER_NAV_MODES,
-	VIEWER_PIN_MODE
 } from '../../../constants/viewer';
 import { UnityUtil } from '../../../globals/unity-util';
-import { getState} from '../../../modules/store';
+import { DialogActions } from '../../../modules/dialog';
+import { dispatch, getState } from '../../../modules/store';
 import { selectMemory } from '../../../modules/viewer';
+import { PIN_COLORS } from '../../../styles';
 import { clientConfigService } from '../../clientConfig';
 import { MultiSelect } from '../multiSelect';
-import { Pin } from '../pin';
-import { UnityController } from './unityController';
 
 const UNITY_LOADER_PATH = 'unity/Build/UnityLoader.js';
 
@@ -38,21 +39,20 @@ interface IViewerConstructor {
 	name?: string;
 }
 
-export class ViewerService extends UnityController {
+export class ViewerService {
 	public element: HTMLElement;
 	public name: string;
 	public viewer: HTMLElement;
-	public previousHighLightedPin = null;
-	public pins = {};
 	public currentNavMode = null;
 	public units = 'm';
 	public convertToM = 1.0;
 	public initialized = false;
 	public measureMode = false;
 	public modelString = null;
-	public pinData = null;
-	private newPinId = 'newPinId';
-	private mode = VIEWER_PIN_MODE.NORMAL;
+	public divId = 'unityViewer';
+	private numClips = 0;
+	private stats: boolean = false;
+	private emitter = new EventEmitter();
 	public initialised: {
 		promise: Promise<any>;
 		resolve: () => void;
@@ -68,7 +68,6 @@ export class ViewerService extends UnityController {
 	public plugins: any;
 
 	public constructor({ name = 'viewer', ...config}: IViewerConstructor) {
-		super();
 		this.name = name;
 
 		this.unityLoaderReady = false;
@@ -88,7 +87,8 @@ export class ViewerService extends UnityController {
 
 	public setupInstance = (container) => {
 		this.element = container;
-		this.startUnity();
+		UnityUtil.init(this.handleUnityError, this.onUnityProgress, this.onModelProgress);
+		UnityUtil.hideProgressBar();
 
 		const unityHolder = document.createElement('div');
 		unityHolder.className = 'emscripten';
@@ -127,7 +127,7 @@ export class ViewerService extends UnityController {
 		}
 
 		this.setInitialisePromise();
-		this.setUnity();
+		UnityUtil.viewer = this;
 
 		try {
 			await this.insertUnityLoader(this.memory);
@@ -197,7 +197,7 @@ export class ViewerService extends UnityController {
 		return new Promise((resolve, reject) => {
 			this.unityLoaderScript.addEventListener ('load', () => {
 				(async () => {
-					await this.loadUnity(memory);
+					await UnityUtil.loadUnity(this.divId, undefined, memory);
 					console.debug('Loaded UnityLoader.js succesfully');
 					resolve();
 				})();
@@ -217,6 +217,118 @@ export class ViewerService extends UnityController {
 	}
 
 	/**
+	 * Emitter
+	 */
+
+	public on = (event, fn, ...args) => {
+		this.emitter.on(event, fn, ...args);
+	}
+
+	public once = (event, fn, ...args) => {
+		this.emitter.once(event, fn, ...args);
+	}
+
+	public off = (event, ...args) => {
+		this.emitter.off(event, ...args);
+	}
+
+	public emit = (event, ...args) => {
+		this.emitter.emit(event, ...args);
+	}
+
+	public removeAllListeners() {
+		this.emitter.removeAllListeners();
+	}
+
+	public pickPointEvent(pointInfo) {
+		// User clicked a mesh
+		this.emit(VIEWER_EVENTS.PICK_POINT, {
+			id : pointInfo.id,
+			normal : pointInfo.normal,
+			position: pointInfo.position,
+			screenPos : pointInfo.mousePos,
+			selectColour : PIN_COLORS.YELLOW,
+		});
+	}
+
+	/**
+	 * Helpers
+	 */
+
+	public async isModelLoaded() {
+		await UnityUtil.onReady();
+		return UnityUtil.onLoaded();
+	}
+
+	public getDefaultHighlightColor() {
+		return UnityUtil.defaultHighlightColor;
+	}
+
+	public getScreenshot() {
+		return UnityUtil.requestScreenShot();
+	}
+
+	public getCurrentViewpointInfo(account, model) {
+		return UnityUtil.requestViewpoint(account, model);
+	}
+
+	public hideHiddenByDefaultObjects() {
+		UnityUtil.hideHiddenByDefaultObjects();
+	}
+
+	public showHiddenByDefaultObjects() {
+		UnityUtil.showHiddenByDefaultObjects();
+	}
+
+	public showCoordView() {
+		UnityUtil.showCoordView();
+	}
+
+	public hideCoordView() {
+		UnityUtil.hideCoordView();
+	}
+
+	public getUnityObjectsStatus(account, model) {
+		return UnityUtil.getObjectsStatus(account, model);
+	}
+
+	/**
+	 * Handlers
+	 */
+
+	private handleUnityError = (message: string, reload: boolean, isUnity: boolean) => {
+		let errorType = '3D Repo Error';
+
+		if (isUnity) {
+			errorType = 'Unity Error';
+		}
+
+		dispatch(DialogActions.showDialog({
+			title: errorType,
+			content: message,
+			onCancel: () => {
+				if (reload) {
+					location.reload();
+				}
+			}
+		}));
+
+		console.error('Unity errored and user canceled reload', message);
+	}
+
+	private onUnityProgress = (progress) => {
+		if (progress === 1) {
+			this.emit(VIEWER_EVENTS.VIEWER_INIT_SUCCESS, progress);
+		} else {
+			this.emit(VIEWER_EVENTS.VIEWER_INIT_PROGRESS, progress);
+		}
+	}
+
+	public onModelProgress = (progress) => {
+		this.emit(VIEWER_EVENTS.MODEL_LOADING_PROGRESS, progress);
+	}
+
+	/**
 	 * Resetting
 	 */
 
@@ -231,7 +343,7 @@ export class ViewerService extends UnityController {
 		// if (this.viewer) {
 		this.setPinDropMode(false);
 		this.initialized = false;
-		this.resetUnity();
+		UnityUtil.reset();
 	// }
 	}
 
@@ -310,21 +422,17 @@ export class ViewerService extends UnityController {
 		return this.initialized && !this.pinDropMode && !this.measureMode;
 	}
 
-	public async highlightObjects(params) {
+	public async highlightObjects(
+		{ account, model, colour, multi, forceReHighlight, ...params }:
+		{ account: string, model: string, colour, multi: boolean, forceReHighlight: boolean, id?: string, ids?: string[]}) {
 		if (this.canHighlight) {
 			const ids = params.id ? [params.id] : params.ids;
 			if (ids) {
 				const uniqueIds = Array.from(new Set(ids));
 				if (uniqueIds.length) {
-					await this.highlightAllObjects(
-							params.account,
-							params.model,
-							ids,
-							params.colour,
-							params.multi,
-							params.forceReHighlightt);
-
-					this.emit(VIEWER_EVENTS.HIGHLIGHT_OBJECTS, {account: params.account, model: params.model, uniqueIds });
+					// @ts-ignore
+					await UnityUtil.highlightObjects(account, model, uniqueIds, colour, multi, forceReHighlight);
+					this.emit(VIEWER_EVENTS.HIGHLIGHT_OBJECTS, {account, model, uniqueIds });
 					return;
 				}
 			}
@@ -333,12 +441,22 @@ export class ViewerService extends UnityController {
 		}
 	}
 
-	public unhighlightObjects(params) {
-		this.unhighlightAllObjects(
-				params.account,
-				params.model,
-				params.id ? [params.id] : params.ids
-		);
+	public clearHighlights() {
+		UnityUtil.clearHighlights();
+		this.emit(VIEWER_EVENTS.CLEAR_HIGHLIGHT_OBJECTS, {});
+	}
+
+	public unhighlightObjects({ account, model, ...params }) {
+		const ids = params.id ? [params.id] : params.ids;
+		if (ids) {
+			const uniqueIds = Array.from(new Set(ids));
+			if (uniqueIds.length) {
+				// @ts-ignore
+				UnityUtil.unhighlightObjects(account, model, uniqueIds);
+				this.emit(VIEWER_EVENTS.UNHIGHLIGHT_OBJECTS, {account, model, uniqueIds });
+				return;
+			}
+		}
 	}
 
 	/**
@@ -370,8 +488,24 @@ export class ViewerService extends UnityController {
 	}
 
 	/**
-	 * Nav mode
+	 * Navigation
 	 */
+
+	public helicopterSpeedDown() {
+		UnityUtil.helicopterSpeedDown();
+	}
+
+	public helicopterSpeedUp() {
+		UnityUtil.helicopterSpeedUp();
+	}
+
+	public helicopterSpeedReset() {
+		UnityUtil.helicopterSpeedReset();
+	}
+
+	public setNavigation(mode) {
+		UnityUtil.setNavigation(mode);
+	}
 
 	public setNavigationMode(mode) {
 		this.setNavMode(mode, false);
@@ -417,51 +551,25 @@ export class ViewerService extends UnityController {
 	 * Pins
 	 */
 
-	public addPin = async ({account, model, id, type, pickedPos: position, norm, colours, viewpoint, isSelected}) => {
-			await this.isViewerReady();
-			this.pins[id] = new Pin({ id, type, position, norm, colors: colours, account, model });
-
-			if (isSelected) {
-				this.selectPin(id);
-			}
-			await this.isModelReady();
-			if (type === 'risk') {
-				await this.dropRiskPin(id, position, norm, colours);
-			} else {
-				await this.dropIssuePin(id, position, norm, colours);
-			}
-	}
-
-	public removePin = ({id}) => {
-		if (this.pins.hasOwnProperty(id)) {
-			this.pins[id].remove();
-			delete this.pins[id];
+	public addPin = async ({ id, type, pickedPos: position, norm, colours, isSelected }) => {
+		await this.isViewerReady();
+		await this.isModelLoaded();
+		if (type === 'risk') {
+			UnityUtil.dropRiskPin(id, position, norm, colours);
+		} else {
+			UnityUtil.dropIssuePin(id, position, norm, colours);
+		}
+		if (isSelected) {
+			UnityUtil.selectPin(id);
 		}
 	}
 
-	public removeUnsavedPin() {
-		this.removePin({ id: this.newPinId });
-		this.setPin({ data: null });
-	}
-
-	public highlightPin(id) {
-		// If a pin was previously highlighted switch it off
-		if (this.previousHighLightedPin) {
-			this.previousHighLightedPin.highlight();
-			this.previousHighLightedPin = null;
-		}
-
-		// If the pin exists switch it on
-		if (id && this.pins.hasOwnProperty(id)) {
-			this.pins[id].highlight();
-			this.previousHighLightedPin = this.pins[id];
-		}
+	public removePin({id}) {
+		UnityUtil.removePin(id);
 	}
 
 	public changePinColor({id, colours}) {
-		if (this.pins.hasOwnProperty(id)) {
-			this.pins[id].changeColour(colours);
-		}
+		UnityUtil.changePinColour(id, colours);
 	}
 
 	public setPinDropMode(on: boolean) {
@@ -472,12 +580,40 @@ export class ViewerService extends UnityController {
 		}
 	}
 
-	public setPin(data) {
-		this.pinData = data;
+	/**
+	 * Diffs
+	 */
+
+	public diffToolSetAsComparator(account: string, model: string) {
+		UnityUtil.diffToolSetAsComparator(account, model);
 	}
 
-	public getPinData(): any {
-		return this.pinData;
+	public diffToolLoadComparator(account: string, model: string, revision: string) {
+		return UnityUtil.diffToolLoadComparator(account, model, revision);
+	}
+
+	public diffToolEnableWithDiffMode() {
+		UnityUtil.diffToolEnableWithDiffMode();
+	}
+
+	public diffToolEnableWithClashMode() {
+		UnityUtil.diffToolEnableWithClashMode();
+	}
+
+	public diffToolDisableAndClear() {
+		UnityUtil.diffToolDisableAndClear();
+	}
+
+	public diffToolShowBaseModel() {
+		UnityUtil.diffToolShowBaseModel();
+	}
+
+	public diffToolShowComparatorModel() {
+		UnityUtil.diffToolShowComparatorModel();
+	}
+
+	public diffToolDiffView() {
+		UnityUtil.diffToolDiffView();
 	}
 
 	/**
@@ -501,6 +637,10 @@ export class ViewerService extends UnityController {
 	 * Load Model
 	 */
 
+	public async isModelReady() {
+		return await this.isModelLoaded();
+	}
+
 	public async loadViewerModel(teamspace, model, branch, revision) {
 		if (!teamspace || !model) {
 			console.error('Teamspace, model, branch or revision was not defined!', teamspace, model, branch, revision);
@@ -516,9 +656,9 @@ export class ViewerService extends UnityController {
 		this.emit(VIEWER_EVENTS.MODEL_LOADING_START);
 		document.body.style.cursor = 'wait';
 
-		await this.loadModel(account, model, branch, revision);
+		await UnityUtil.loadModel(account, model, branch, revision);
 
-		await this.onLoaded().then((bbox) => {
+		await UnityUtil.onLoaded().then((bbox) => {
 			document.body.style.cursor = 'initial';
 
 			this.emit(VIEWER_EVENTS.MODEL_LOADED, 1);
@@ -530,29 +670,36 @@ export class ViewerService extends UnityController {
 			}
 		});
 
-		return this.onLoading();
-	}
-
-	/**
-	 * Model Ready
-	 */
-
-	public async isModelReady() {
-		return await this.isModelLoaded();
+		return UnityUtil.onLoading();
 	}
 
 	/**
 	 * Zooms
 	 */
 
-	public async zoomToHighlightedMeshes() {
-		await this.isModelReady();
-		super.zoomToHighlightedMeshes();
+	public zoomToHighlightedMeshes() {
+		UnityUtil.zoomToHighlightedMeshes();
 	}
 
-	public async zoomToObjects(meshes) {
-		await this.isModelReady();
-		super.zoomToObjects(meshes);
+	public zoomToObjects(meshes) {
+		UnityUtil.zoomToObjects(meshes);
+	}
+
+	/**
+	 * Mesh Color
+	 */
+
+	public overrideMeshColor(account, model, meshIDs, color) {
+		UnityUtil.overrideMeshColor(account, model, meshIDs, color);
+
+		if (color.length > 3) {
+			UnityUtil.overrideMeshOpacity(account, model, meshIDs, color[3]);
+		}
+	}
+
+	public resetMeshColor(account, model, meshIDs) {
+		UnityUtil.resetMeshOpacity(account, model, meshIDs);
+		UnityUtil.resetMeshColor(account, model, meshIDs);
 	}
 
 	/**
@@ -562,10 +709,10 @@ export class ViewerService extends UnityController {
 	public setFarPlaneAlgorithm = (algorithm: string) => {
 		switch (algorithm) {
 			case 'box':
-				this.useBoundingBoxFarPlaneAlgorithm();
+				UnityUtil.useBoundingBoxFarPlaneAlgorithm();
 				break;
 			case 'sphere':
-				this.useBoundingSphereFarPlaneAlgorithm();
+				UnityUtil.useBoundingSphereFarPlaneAlgorithm();
 				break;
 		}
 	}
@@ -573,19 +720,46 @@ export class ViewerService extends UnityController {
 	public setShading = (shading: string) => {
 		switch (shading) {
 			case 'standard':
-				this.setRenderingQualityDefault();
+				UnityUtil.setRenderingQualityDefault();
 				break;
 			case 'architectural':
-				this.setRenderingQualityHigh();
+				UnityUtil.setRenderingQualityHigh();
 				break;
 		}
 	}
 
 	public setXray = (xray: boolean) => {
 		if (xray) {
-			this.setXRayHighlightOn();
+			UnityUtil.setXRayHighlightOn();
 		} else {
-			this.setXRayHighlightOff();
+			UnityUtil.setXRayHighlightOff();
+		}
+	}
+
+	public setMaxShadowDistance(value: number) {
+		if (value === undefined) { return; }
+		UnityUtil.setMaxShadowDistance(value);
+	}
+
+	public setNumCacheThreads(value: number) {
+		if (value === undefined) { return; }
+		UnityUtil.setNumCacheThreads(value);
+	}
+
+	public setNearPlane = (nearplane: number) => {
+		if (nearplane === undefined) { return; }
+		UnityUtil.setDefaultNearPlane(nearplane);
+	}
+
+	public setFarPlaneSamplingPoints = (farplaneSample: number) => {
+		if (farplaneSample === undefined) { return; }
+		UnityUtil.setFarPlaneSampleSize(farplaneSample);
+	}
+
+	public setStats = (val: boolean = false) => {
+		if (val !== this.stats) {
+			UnityUtil.toggleStats();
+			this.stats = val;
 		}
 	}
 
@@ -595,32 +769,32 @@ export class ViewerService extends UnityController {
 
 	public async resetMapSources() {
 		await this.isViewerReady();
-		super.resetMapSources();
+		UnityUtil.resetMapSources();
 	}
 
 	public async addMapSource(source) {
 		await this.isViewerReady();
-		super.addMapSource(source);
+		UnityUtil.addMapSource(source);
 	}
 
 	public async removeMapSource(source) {
 		await this.isViewerReady();
-		super.removeMapSource(source);
+		UnityUtil.removeMapSource(source);
 	}
 
 	public async mapInitialise(surveyPoints) {
 		await this.isViewerReady();
-		super.mapInitialise(surveyPoints);
+		UnityUtil.mapInitialise(surveyPoints);
 	}
 
 	public async mapStart() {
 		await this.isViewerReady();
-		super.mapStart();
+		UnityUtil.mapStart();
 	}
 
 	public async mapStop() {
 		await this.isViewerReady();
-		super.mapStop();
+		UnityUtil.mapStop();
 	}
 
 	/**
@@ -650,12 +824,148 @@ export class ViewerService extends UnityController {
 		return this.showAll();
 	}
 
-	public async setCamera({ position, up, view_dir, look_at, animate, rollerCoasterMode }) {
+	public async setCamera({ position, up, view_dir, look_at, animate, rollerCoasterMode, account, model }) {
 		await this.isModelReady();
+		UnityUtil.setViewpoint(position, up, view_dir, look_at, animate !== undefined ? animate : true, rollerCoasterMode);
+	}
 
-		return this.setCameraPosition(
-			position, up, view_dir, look_at, animate !== undefined ? animate : true, rollerCoasterMode
-		);
+	/**
+	 * Camera
+	 */
+
+	public topView() {
+		UnityUtil.topView();
+	}
+
+	public bottomView() {
+		return UnityUtil.bottomView();
+	}
+
+	public frontView() {
+		return UnityUtil.frontView();
+	}
+
+	public backView() {
+		return UnityUtil.backView();
+	}
+
+	public leftView() {
+		return UnityUtil.leftView();
+	}
+
+	public rightView() {
+		return UnityUtil.rightView();
+	}
+
+	public showAll() {
+		UnityUtil.resetCamera();
+	}
+
+	public centreToPoint(params) {
+		UnityUtil.centreToPoint(params);
+	}
+
+	/**
+	 * Controlling
+	 */
+
+	public pauseRendering() {
+		UnityUtil.pauseRendering();
+	}
+
+	public resumeRendering() {
+		UnityUtil.resumeRendering();
+	}
+
+	public setModelCache = (cache: boolean) => {
+		if (cache) {
+			UnityUtil.enableCaching();
+		} else {
+			UnityUtil.disableCaching();
+		}
+	}
+
+	public switchObjectVisibility(account, model, ids, visibility) {
+		UnityUtil.toggleVisibility(account, model, ids, visibility);
+	}
+
+	public startAreaSelect() {
+		UnityUtil.startAreaSelection();
+	}
+
+	public stopAreaSelect() {
+		UnityUtil.stopAreaSelection();
+	}
+
+	public setShadows = (type: string) => {
+		switch (type) {
+			case 'soft':
+				this.enableSoftShadows();
+				break;
+			case 'hard':
+				this.enableHardShadows();
+				break;
+			case 'none':
+				this.disableShadows();
+				break;
+		}
+	}
+
+	public enableSoftShadows() {
+		UnityUtil.enableSoftShadows();
+	}
+
+	public enableHardShadows() {
+		UnityUtil.enableHardShadows();
+	}
+
+	public disableShadows() {
+		UnityUtil.disableShadows();
+	}
+
+	/**
+	 * Clip
+	 */
+
+	public getNumPlanes() {
+		return this.numClips;
+	}
+
+	public clipBroadcast(clip) {
+		this.emit(VIEWER_EVENTS.CLIPPING_PLANE_BROADCAST, clip);
+	}
+
+	public numClipPlanesUpdated(nPlanes) {
+		this.numClips = nPlanes;
+		this.emit(VIEWER_EVENTS.UPDATE_NUM_CLIP, nPlanes);
+	}
+
+	public updateClippingPlanes({ clipPlanes, account, model }: any) {
+		UnityUtil.updateClippingPlanes(clipPlanes ? clipPlanes : [], false, account, model);
+	}
+
+	public startClip(isSingle: boolean) {
+		if (isSingle) {
+			this.startSingleClip();
+		} else {
+			this.startBoxClip();
+		}
+	}
+
+	public startBoxClip() {
+		UnityUtil.startBoxClip();
+	}
+
+	public startSingleClip() {
+		UnityUtil.startSingleClip();
+	}
+
+	public startClipEdit() {
+		UnityUtil.startClipEdit();
+	}
+
+	public stopClipEdit() {
+		UnityUtil.stopClipEdit();
 	}
 }
